+++ src/TvDbSharper/Dto/UpdatesDtos.cs
@@ -1,5 +1,10 @@
-﻿namespace TvDbSharper.Dto
+﻿using Newtonsoft.Json;
+
+namespace TvDbSharper.Dto
 {
+#if DEBUG
+    [JsonObject(ItemRequired = Required.AllowNull)]
+#endif
     public class Update
     {
         public int Id { get; set; }