+++ src/TvDbSharper/ISeriesClient.cs
@@ -189,6 +189,25 @@
         /// <param name="query">The query.</param>
         /// <returns>Returns <see cref="T:System.Threading.Tasks.Task`1" />.The task object representing the asynchronous operation.</returns>
         Task<TvDbResponse<Image[]>> GetImagesAsync(int seriesId, ImagesQuery query);
+        
+        /// <summary>
+        /// <para>[GET /series/{id}/images/query]</para>
+        /// <para>Query images for the given series ID.</para>
+        /// </summary>
+        /// <param name="seriesId">The series ID</param>
+        /// <param name="query">The query.</param>
+        /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+        /// <returns>Returns <see cref="T:System.Threading.Tasks.Task`1" />.The task object representing the asynchronous operation.</returns>
+        Task<TvDbResponse<Image[]>> GetImagesAsync(int seriesId, ImagesQueryAlternative query, CancellationToken cancellationToken);
+
+        /// <summary>
+        /// <para>[GET /series/{id}/images/query]</para>
+        /// <para>Query images for the given series ID.</para>
+        /// </summary>
+        /// <param name="seriesId">The series ID</param>
+        /// <param name="query">The query.</param>
+        /// <returns>Returns <see cref="T:System.Threading.Tasks.Task`1" />.The task object representing the asynchronous operation.</returns>
+        Task<TvDbResponse<Image[]>> GetImagesAsync(int seriesId, ImagesQueryAlternative query);
 
         /// <summary>
         /// <para>[GET /series/{id}/images]</para>